Add test for removing function
